script to open google calendar page with firefox
This script should be called from the qtile GoogleCalendar
widget.  Set browser_cmd='open_calendar.py' (with appropriate
path prefix).  The script will parse sessionstore.js to
see if a Google Calendar page is open.  If so, it will do
nothing.  If not, it will open a page to the Google Calendar.